requirements.txt
google-adk
fastapi
uvicorn
google-cloud-firestore
google-cloud-aiplatform
google-cloud-logging
python-dotenv
streamlit
requests